refactor with src directory
