+++ apps/web-app/src/pages/index.tsx
@@ -1,5 +1,5 @@
 import {sayHello} from '@optional-package-scope/foo'
-import {AsyncBar, Bar} from '@optional-package-scope/bar'
+import {AsyncMessage, Message} from '@optional-package-scope/bar'
 
 export default function Home() {
   return (
@@ -7,8 +7,8 @@ export default function Home() {
       Imported and transpiled modules from other workspaces:
       <ul>
           <li>{`Foo says: ${sayHello('World')} from @optional-package-scope/foo`}</li>
-          <li><Bar message={'Bar react component from @optional-package-scope/bar'}/></li>
-          <li><AsyncBar apiUrl={'/api/hello'}/></li>
+          <li><Message message={'Bar react component from @optional-package-scope/bar'}/></li>
+          <li><AsyncMessage apiUrl={'/api/hello'}/></li>
       </ul>
     </div>
   )
