+++ packages/bar/src/index.ts
@@ -1,0 +1,2 @@
+export {Message} from './message';
+export {AsyncMessage} from './async-message';
