+++ packages/bar/src/message.tsx
@@ -1,0 +1,6 @@
+type Props = {
+    message: string;
+}
+
+export const Message = (props: Props) => <span>{props.message}</span>;
+
