+++ packages/bar/tsconfig.json
@@ -19,7 +19,8 @@
     "jsx": "preserve"
   },
   "exclude": [
-    "node_modules"
+    "node_modules",
+    "dist"
   ],
   "include": [
     "**/*.ts",
