+++ packages/foo/index.ts
@@ -1,4 +1,1 @@
-export const sayHello = (name: string) => {
-    return `Hello ${name} !`
-}
-
+export * from './src/index'
