+++ packages/foo/src/index.ts
@@ -1,0 +1,4 @@
+export const sayHello = (name: string) => {
+    return `Hello ${name} !`
+}
+
